# server/internal/rest/swagger/swagger.yaml
basePath: /
definitions:
  main.CreateUserRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      username:
        example: vlad
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  main.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
  main.MessageResponse:
    properties:
      message:
        example: pong
        type: string
    type: object
  main.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      username:
        example: vlad
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для приложения вишлистов.
  title: Wishlist API
  version: "1.0"
paths:
  /api/v1/echo:
    post:
      consumes:
      - application/json
      description: Принимает произвольный JSON и возвращает его обратно.
      parameters:
      - description: Любой JSON
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Вернуть переданный JSON
      tags:
      - debug
  /api/v1/ping:
    get:
      description: Возвращает простой JSON-ответ.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.MessageResponse'
      summary: Проверка доступности API
      tags:
      - system
  /api/v1/users:
    post:
      consumes:
      - application/json
      description: Принимает JSON, биндингует его в Go-структуру и возвращает созданного
        пользователя.
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Создать пользователя
      tags:
      - users
  /api/v1/users/{id}:
    get:
      description: Возвращает пользователя по идентификатору.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Получить пользователя по ID
      tags:
      - users
swagger: "2.0"